feat!: Splits borsh into borsh and borsh_annotation for better dependency management.
