+++ packages/borsh/lib/borsh.dart
@@ -1,7 +1,1 @@
 export 'package:borsh/src/extensions.dart';
-export 'package:borsh/src/field_annotation.dart'
-    show u8, i8, u16, i16, u32, i32, u64, i64, string, Array;
-export 'package:borsh/src/field_annotation.dart';
-export 'package:borsh/src/struct.dart' show BorshStruct, EmptyBorshStruct;
-export 'package:borsh/src/struct_annotation.dart' show Struct, IOExt;
-export 'package:borsh/src/type.dart' show Borsh;
