+++ packages/borsh/lib/src/field_info_generators/size_calculation.dart
@@ -1,7 +1,7 @@
 import 'package:borsh/src/extensions.dart';
 import 'package:borsh/src/field_info_generators/utils.dart';
 import 'package:borsh/src/generator_utils.dart';
-import 'package:borsh/src/type.dart';
+import 'package:borsh_annotation/borsh_annotation.dart';
 
 String generateSizeComputationCode(List<FieldInfo> fields) {
   final StringBuffer buffer = StringBuffer();
