+++ packages/borsh/pubspec.yaml
@@ -1,5 +1,5 @@
 name: borsh
-version: 0.1.0
+version: 0.2.0
 repository: https://github.com/cryptoplease/cryptoplease-dart
 description: >-
   Implementation of the borsh serialization scheme
@@ -9,6 +9,7 @@ environment:
 
 dependencies:
   analyzer: ^2.0.0
+  borsh_annotation: ^0.2.0
   build: ^2.0.3
   source_gen: ^1.0.3
