+++ packages/borsh/test/borsh_test_structs/simple_struct1.dart
@@ -1,6 +1,7 @@
 import 'dart:typed_data';
 
-import 'package:borsh/borsh.dart';
+import 'package:borsh_annotation/borsh_annotation.dart';
+import 'package:borsh_annotation/src/struct.dart';
 
 part 'simple_struct1.g.dart';
