+++ packages/borsh_annotation/CHANGELOG.md
@@ -1,0 +1,4 @@
+## 0.1.0
+
+ - Initial version
+
