+++ packages/borsh_annotation/README.md
@@ -1,0 +1,1 @@
+# Basic library for borsh serialization
