+++ packages/borsh_annotation/build.yaml
@@ -1,0 +1,19 @@
+targets:
+  $default:
+    builders:
+      borsh|borsh:
+        enabled: true
+
+builders:
+  borsh:
+    target: ":borsh"
+    import: "package:borsh/src/builder.dart"
+    builder_factories:
+      - "borsh"
+    build_extensions:
+      ".dart":
+        - ".borsh.g.part"
+    auto_apply: "dependents"
+    build_to: "cache"
+    applies_builders:
+      - "source_gen|combining_builder"
