+++ packages/borsh_annotation/lib/borsh_annotation.dart
@@ -1,0 +1,4 @@
+export 'package:borsh_annotation/src/field_annotation.dart';
+export 'package:borsh_annotation/src/struct.dart';
+export 'package:borsh_annotation/src/struct_annotation.dart' show Struct, IOExt;
+export 'package:borsh_annotation/src/type.dart';
