+++ packages/borsh_annotation/lib/src/struct.dart
